add validation on log in and sign up file and change date formate

+++ src/screen/Auth/Login/Login.jsx
@@ -8,52 +8,69 @@ import { fetchUserData } from "../../../service/AuthService";
 import Swal from "sweetalert2";
 import { setItem } from "../../../helper/Storage";
 import authContext from "../../../context/AuthContext";
+import { checkBlankUserInput } from "../../../helper/Validation";
 
 const Login = () => {
   const { onLogin } = useContext(authContext);
   const navigator = useNavigate();
   const [userEmail, setUserEmail] = useState("");
   const [userPassword, setUserPassword] = useState("");
+  const [emailValidate, setEmailValidate] = useState("");
+  const [passWordValidate, setPassWordValidate] = useState("");
   const [error, setError] = useState();
 
   const userEmailHandler = (value) => {
     setUserEmail(value);
     setError();
+    setEmailValidate("");
   };
   const userPasswordHandler = (value) => {
     setUserPassword(value);
     setError();
+    setPassWordValidate("");
   };
 
   const handlerSubmit = async (e) => {
+
     e.preventDefault();
-    const userData = {
-      email: userEmail,
-      password: userPassword,
-      returnSecureToken: true,
-    };
-    const response = await fetchUserData(userData);
-    console.log(response);
-    if (response?.data?.idToken) {
-      Swal.fire({
-        position: "center-center",
-        icon: "success",
-        title: `Welcome back ${response?.data.displayName}`,
-        showConfirmButton: false,
-        timer: 1500,
-      }).then(() => {
-        onLogin(true);
-        setItem("user", {
-          userToken: response?.data?.idToken,
-          email: response?.data?.email,
-          userName: response?.data.displayName,
-          userId: response?.data.localId,
-        });
-        navigator("/");
-      });
-    } else {
-      setError(response?.response?.data?.error.message);
+    const isEmailVaild    = checkBlankUserInput(userEmail);
+    const isPassWordVaild = checkBlankUserInput(userPassword);
+    if(isEmailVaild) {
+      setEmailValidate("Please enter email ");
     }
+    if(isPassWordVaild){
+      setPassWordValidate("Please enter password");
+    }
+    if(userEmail!=="" && userPassword !== "") {
+      const userData = {
+        email: userEmail,
+        password: userPassword,
+        returnSecureToken: true,
+      };
+      const response = await fetchUserData(userData);
+      if (response?.data?.idToken) {
+        Swal.fire({
+          position: "center-center",
+          icon: "success",
+          title: `Welcome back ${response?.data.displayName}`,
+          showConfirmButton: false,
+          timer: 1500,
+        }).then(() => {
+          onLogin(true);
+          setItem("user", {
+            userToken: response?.data?.idToken,
+            email: response?.data?.email,
+            userName: response?.data.displayName,
+            userId: response?.data.localId,
+          });
+          navigator("/");
+        });
+      } else {
+        const errorMessage = response?.response?.data?.error.message.charAt(0).toUpperCase() + response?.response?.data?.error.message.slice(1).toLowerCase();
+        setError(() => errorMessage.replace("_", " "));
+      }
+    }
+   
   };
 
   return (
@@ -68,6 +85,7 @@ const Login = () => {
             value={userEmail}
             name="userName"
           />
+          <span style={{ color: "red" }}>{emailValidate}</span>
         </div>
         <div className="mb-3">
           <label className="form-label">password</label>
@@ -78,10 +96,11 @@ const Login = () => {
             onChange={(e) => userPasswordHandler(e.target.value)}
             placeholder="Password"
           />
+          <span style={{ color: "red" }}>{passWordValidate}</span>
           {error && <p style={{ color: "red" }}>{error}</p>}
         </div>
         <div className="mt-3">
-          <Button type="submit" classes="form-control authButton" disable = {userEmail === "" || userPassword === ""? true : false}>
+          <Button type="submit" classes="form-control authButton">
             Login
           </Button>
         </div>
